cleanup and warnings eliminated

+++ src/include/libmspdriver/hw_memmap.h
@@ -46,6 +46,7 @@
 
 #include "stdint.h"
 #include "stdbool.h"
+#include "stddef.h"
 
 //*****************************************************************************
 //
@@ -68,10 +69,10 @@
 //
 //*****************************************************************************
 #define HWREG32(x)                                                              \
-    (*((volatile uint32_t *)((uint16_t)x)))
+    (*((volatile uint32_t *)((size_t)x)))
 #define HWREG16(x)                                                             \
-    (*((volatile uint16_t *)((uint16_t)x)))
+    (*((volatile uint16_t *)((size_t)x)))
 #define HWREG8(x)                                                             \
-    (*((volatile uint8_t *)((uint16_t)x)))
+    (*((volatile uint8_t *)((size_t)x)))
 
 #endif // #ifndef __HW_MEMMAP__
